move vercel constants to common folder to fix lint check
