+++ components/vercel_token_auth/vercel_token_auth.app.mjs
@@ -1,5 +1,5 @@
 import { axios } from "@pipedream/platform";
-import constants from "./constants.mjs";
+import constants from "./common/constants.mjs";
 
 export default {
   type: "app",
